+++ packages/calcite-components/src/components/block/block.stories.ts
@@ -303,3 +303,29 @@ export const toggleDisplayWithLongText_TestOnly = (): string =>
       </calcite-notice>
     </calcite-block-section>
   </calcite-block>`;
+
+export const icons_TestOnly = (): string => html`
+  <calcite-block heading="Heading" description="summary" collapsible open>
+    <calcite-block-section
+      text="Planes, trains, and automobiles are some examples of modes of transportation"
+      open
+      icon-end="pen"
+      icon-start="pen"
+      toggle-display="switch"
+      status="valid"
+    >
+      <p>Block section content</p>
+    </calcite-block-section>
+
+    <calcite-block-section
+      text="Planes, trains, and automobiles are some examples of modes of transportation"
+      open
+      icon-end="pen"
+      icon-start="pen"
+      toggle-display="button"
+      status="valid"
+    >
+      <p>Block section content</p>
+    </calcite-block-section>
+  </calcite-block>
+`;
